Rename _utils.py to thorlabs_utils.py
